Remove use client in page.tsx.
