+++ constants/index.ts
@@ -29,3 +29,6 @@ export const avatarPlaceholderUrl =
 
 export const avatarUserPlaceholder =
   "https://i.pinimg.com/736x/e9/e0/7d/e9e07de22e3ef161bf92d1bcf241e4d0.jpg";
+
+export const postImagePlaceholder =
+  "https://i.pinimg.com/736x/7e/6a/7a/7e6a7ad6ede2b31f94dc38a6fcd7752e.jpg";
